+++ src/ESP32_PWM.hpp
@@ -21,7 +21,7 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.2.1
+  Version: 1.2.2
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -31,6 +31,7 @@
   1.1.1   K Hoang      09/11/2021 Fix examples to not use GPIO1/TX0 for core v2.0.1+
   1.2.0   K Hoang      29/01/2022 Fix multiple-definitions linker error. Improve accuracy. Fix bug
   1.2.1   K Hoang      30/01/2022 DutyCycle to be updated at the end current PWM period
+  1.2.2   K Hoang      01/02/2022 Use float for DutyCycle and Freq, uint32_t for period. Optimize code
 *****************************************************************************************************************************/
 
 #pragma once
@@ -63,13 +64,13 @@
 #endif
 
 #ifndef ESP32_PWM_VERSION
-  #define ESP32_PWM_VERSION           "ESP32_PWM v1.2.1"
+  #define ESP32_PWM_VERSION           "ESP32_PWM v1.2.2"
   
   #define ESP32_PWM_VERSION_MAJOR     1
   #define ESP32_PWM_VERSION_MINOR     2
-  #define ESP32_PWM_VERSION_PATCH     1
+  #define ESP32_PWM_VERSION_PATCH     2
 
-  #define ESP32_PWM_VERSION_INT       1002001
+  #define ESP32_PWM_VERSION_INT       1002002
 #endif
 
 #ifndef TIMER_INTERRUPT_DEBUG