+++ src/ESP32_PWM_ISR.hpp
@@ -20,7 +20,7 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.2.1
+  Version: 1.2.2
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -30,6 +30,7 @@
   1.1.1   K Hoang      09/11/2021 Fix examples to not use GPIO1/TX0 for core v2.0.1+
   1.2.0   K Hoang      29/01/2022 Fix multiple-definitions linker error. Improve accuracy. Fix bug
   1.2.1   K Hoang      30/01/2022 DutyCycle to be updated at the end current PWM period
+  1.2.2   K Hoang      01/02/2022 Use float for DutyCycle and Freq, uint32_t for period. Optimize code
 *****************************************************************************************************************************/
 
 #pragma once
@@ -51,13 +52,13 @@
 #endif
 
 #ifndef ESP32_PWM_VERSION
-  #define ESP32_PWM_VERSION           "ESP32_PWM v1.2.1"
+  #define ESP32_PWM_VERSION           "ESP32_PWM v1.2.2"
   
   #define ESP32_PWM_VERSION_MAJOR     1
   #define ESP32_PWM_VERSION_MINOR     2
-  #define ESP32_PWM_VERSION_PATCH     1
+  #define ESP32_PWM_VERSION_PATCH     2
 
-  #define ESP32_PWM_VERSION_INT       1002001
+  #define ESP32_PWM_VERSION_INT       1002002
 #endif
 
 #ifndef _PWM_LOGLEVEL_
@@ -114,10 +115,10 @@ class ESP32_PWM_ISR
     //////////////////////////////////////////////////////////////////
     // PWM
     // Return the channelNum if OK, -1 if error
-    int setPWM(const uint32_t& pin, const double& frequency, const double& dutycycle, timer_callback StartCallback = nullptr, 
+    int setPWM(const uint32_t& pin, const float& frequency, const float& dutycycle, timer_callback StartCallback = nullptr, 
                 timer_callback StopCallback = nullptr)
     {
-      double period = 0.0;
+      uint32_t period = 0;
       
       if ( ( frequency > 0.0 ) && ( frequency <= 500.0 ) )
       {
@@ -141,8 +142,8 @@ class ESP32_PWM_ISR
 
     // period in us
     // Return the channelNum if OK, -1 if error
-    int setPWM_Period(const uint32_t& pin, const double& period, const double& dutycycle, timer_callback StartCallback = nullptr,
-                       timer_callback StopCallback = nullptr)  
+    int setPWM_Period(const uint32_t& pin, const uint32_t& period, const float& dutycycle, 
+                      timer_callback StartCallback = nullptr, timer_callback StopCallback = nullptr)  
     {     
       return setupPWMChannel(pin, period, dutycycle, (void *) StartCallback, (void *) StopCallback);      
     } 
@@ -151,9 +152,9 @@ class ESP32_PWM_ISR
     
     // low level function to modify a PWM channel
     // returns the true on success or false on failure
-    bool modifyPWMChannel(const unsigned& channelNum, const uint32_t& pin, const double& frequency, const double& dutycycle)
+    bool modifyPWMChannel(const uint8_t& channelNum, const uint32_t& pin, const float& frequency, const float& dutycycle)
     {
-      double period = 0.0;
+      uint32_t period = 0;
       
       if ( ( frequency > 0.0 ) && ( frequency <= 500.0 ) )
       {
@@ -175,22 +176,22 @@ class ESP32_PWM_ISR
     }
     
     //period in us
-    bool modifyPWMChannel_Period(const unsigned& channelNum, const uint32_t& pin, const double& period, const double& dutycycle);
+    bool modifyPWMChannel_Period(const uint8_t& channelNum, const uint32_t& pin, const uint32_t& period, const float& dutycycle);
 
     // destroy the specified PWM channel
-    void deleteChannel(const unsigned& channelNum);
+    void deleteChannel(const uint8_t& channelNum);
 
     // restart the specified PWM channel
-    void restartChannel(const unsigned& channelNum);
+    void restartChannel(const uint8_t& channelNum);
 
     // returns true if the specified PWM channel is enabled
-    bool isEnabled(const unsigned& channelNum);
+    bool isEnabled(const uint8_t& channelNum);
 
     // enables the specified PWM channel
-    void enable(const unsigned& channelNum);
+    void enable(const uint8_t& channelNum);
 
     // disables the specified PWM channel
-    void disable(const unsigned& channelNum);
+    void disable(const uint8_t& channelNum);
 
     // enables all PWM channels
     void enableAll();
@@ -199,13 +200,13 @@ class ESP32_PWM_ISR
     void disableAll();
 
     // enables the specified PWM channel if it's currently disabled, and vice-versa
-    void toggle(const unsigned& channelNum);
+    void toggle(const uint8_t& channelNum);
 
     // returns the number of used PWM channels
-    unsigned getnumChannels();
+    int8_t getnumChannels();
 
     // returns the number of available PWM channels
-    unsigned getNumAvailablePWMChannels() 
+    uint8_t getNumAvailablePWMChannels() 
     {
       return MAX_NUMBER_CHANNELS - numChannels;
     };
@@ -215,7 +216,7 @@ class ESP32_PWM_ISR
     // low level function to initialize and enable a new PWM channel
     // returns the PWM channel number (channelNum) on success or
     // -1 on failure (f == NULL) or no free PWM channels 
-    int setupPWMChannel(const uint32_t& pin, const double& period, const double& dutycycle, void* cbStartFunc = nullptr, void* cbStopFunc = nullptr);
+    int setupPWMChannel(const uint32_t& pin, const uint32_t& period, const float& dutycycle, void* cbStartFunc = nullptr, void* cbStopFunc = nullptr);
 
     // find the first available slot
     int findFirstFreeSlot();
@@ -228,7 +229,7 @@ class ESP32_PWM_ISR
       ///////////////////////////////////
       
       uint64_t      prevTime;           // value returned by the micros() or millis() function in the previous run() call
-      double        period;             // period value, in us / ms
+      uint32_t      period;             // period value, in us / ms
       uint32_t      onTime;             // onTime value, ( period * dutyCycle / 100 ) us  / ms
       
       void*         callbackStart;      // pointer to the callback function when PWM pulse starts (HIGH)
@@ -243,15 +244,16 @@ class ESP32_PWM_ISR
       bool          enabled;            // true if enabled
       
       // New from v1.2.1   
-      double        newPeriod;          // period value, in us / ms
-      double        newDutyCycle;       // from 0.00 to 100.00, double precision
+      uint32_t      newPeriod;          // period value, in us / ms
+      uint32_t      newOnTime;          // onTime value, ( period * dutyCycle / 100 ) us  / ms
+      float         newDutyCycle;       // from 0.00 to 100.00, float precision
       //////
     } PWM_t;
 
     volatile PWM_t PWM[MAX_NUMBER_CHANNELS];
 
     // actual number of PWM channels in use (-1 means uninitialized)
-    volatile int numChannels;
+    volatile int8_t numChannels;
 
     // ESP32 is a multi core / multi processing chip. It is mandatory to disable task switches during ISR
     portMUX_TYPE PWM_Mux = portMUX_INITIALIZER_UNLOCKED;