+++ src/ESP32_PWM_ISR_Impl.h
@@ -20,7 +20,7 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.2.1
+  Version: 1.2.2
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -30,6 +30,7 @@
   1.1.1   K Hoang      09/11/2021 Fix examples to not use GPIO1/TX0 for core v2.0.1+
   1.2.0   K Hoang      29/01/2022 Fix multiple-definitions linker error. Improve accuracy. Fix bug
   1.2.1   K Hoang      30/01/2022 DutyCycle to be updated at the end current PWM period
+  1.2.2   K Hoang      01/02/2022 Use float for DutyCycle and Freq, uint32_t for period. Optimize code
 *****************************************************************************************************************************/
 
 #pragma once
@@ -133,7 +134,7 @@ void IRAM_ATTR ESP32_PWM_ISR::run()
           PWM[channelNum].period    = PWM[channelNum].newPeriod;
           PWM[channelNum].newPeriod = 0;
           
-          PWM[channelNum].onTime  = ( PWM[channelNum].period * PWM[channelNum].newDutyCycle ) / 100;
+          PWM[channelNum].onTime  = PWM[channelNum].newOnTime;
         }
 #endif
       }        
@@ -172,12 +173,12 @@ int ESP32_PWM_ISR::findFirstFreeSlot()
 
 ///////////////////////////////////////////////////
 
-int ESP32_PWM_ISR::setupPWMChannel(const uint32_t& pin, const double& period, const double& dutycycle, void* cbStartFunc, void* cbStopFunc)
+int ESP32_PWM_ISR::setupPWMChannel(const uint32_t& pin, const uint32_t& period, const float& dutycycle, void* cbStartFunc, void* cbStopFunc)
 {
   int channelNum;
   
   // Invalid input, such as period = 0, etc
-  if ( (period <= 0.0) || (dutycycle < 0.0) || (dutycycle > 100.0) )
+  if ( (period == 0) || (dutycycle < 0.0) || (dutycycle > 100.0) )
   {
     PWM_LOGERROR("Error: Invalid period or dutycycle");
     return -1;
@@ -232,10 +233,10 @@ int ESP32_PWM_ISR::setupPWMChannel(const uint32_t& pin, const double& period, co
 
 ///////////////////////////////////////////////////
 
-bool ESP32_PWM_ISR::modifyPWMChannel_Period(const unsigned& channelNum, const uint32_t& pin, const double& period, const double& dutycycle)
+bool ESP32_PWM_ISR::modifyPWMChannel_Period(const uint8_t& channelNum, const uint32_t& pin, const uint32_t& period, const float& dutycycle)
 {
   // Invalid input, such as period = 0, etc
-  if ( (period <= 0.0) || (dutycycle < 0.0) || (dutycycle > 100.0) )
+  if ( (period == 0) || (dutycycle < 0.0) || (dutycycle > 100.0) )
   {
     PWM_LOGERROR("Error: Invalid period or dutycycle");
     return false;
@@ -260,13 +261,14 @@ bool ESP32_PWM_ISR::modifyPWMChannel_Period(const unsigned& channelNum, const ui
 
   PWM[channelNum].newPeriod     = period;
   PWM[channelNum].newDutyCycle  = dutycycle;
+  PWM[channelNum].newOnTime     = ( period * dutycycle ) / 100;
   
   // ESP32 is a multi core / multi processing chip. It is mandatory to disable task switches during modifying shared vars
   portEXIT_CRITICAL(&PWM_Mux);
   
   PWM_LOGDEBUG0("Channel : ");      PWM_LOGDEBUG0(channelNum); 
   PWM_LOGDEBUG0("\tNew Period : "); PWM_LOGDEBUG0(PWM[channelNum].newPeriod);
-  PWM_LOGDEBUG0("\t\tOnTime : ");   PWM_LOGDEBUG0(( period * dutycycle ) / 100); 
+  PWM_LOGDEBUG0("\t\tOnTime : ");   PWM_LOGDEBUG0(PWM[channelNum].newOnTime); 
   PWM_LOGDEBUG0("\tStart_Time : "); PWM_LOGDEBUGLN0(PWM[channelNum].prevTime);
   
 #else
@@ -295,7 +297,7 @@ bool ESP32_PWM_ISR::modifyPWMChannel_Period(const unsigned& channelNum, const ui
 
 ///////////////////////////////////////////////////
 
-void ESP32_PWM_ISR::deleteChannel(const unsigned& channelNum) 
+void ESP32_PWM_ISR::deleteChannel(const uint8_t& channelNum) 
 {
   if (channelNum >= MAX_NUMBER_CHANNELS) 
   {
@@ -330,7 +332,7 @@ void ESP32_PWM_ISR::deleteChannel(const unsigned& channelNum)
 ///////////////////////////////////////////////////
 
 // function contributed by code@rowansimms.com
-void ESP32_PWM_ISR::restartChannel(const unsigned& channelNum) 
+void ESP32_PWM_ISR::restartChannel(const uint8_t& channelNum) 
 {
   if (channelNum >= MAX_NUMBER_CHANNELS) 
   {
@@ -348,7 +350,7 @@ void ESP32_PWM_ISR::restartChannel(const unsigned& channelNum)
 
 ///////////////////////////////////////////////////
 
-bool ESP32_PWM_ISR::isEnabled(const unsigned& channelNum) 
+bool ESP32_PWM_ISR::isEnabled(const uint8_t& channelNum) 
 {
   if (channelNum >= MAX_NUMBER_CHANNELS) 
   {
@@ -360,7 +362,7 @@ bool ESP32_PWM_ISR::isEnabled(const unsigned& channelNum)
 
 ///////////////////////////////////////////////////
 
-void ESP32_PWM_ISR::enable(const unsigned& channelNum) 
+void ESP32_PWM_ISR::enable(const uint8_t& channelNum) 
 {
   if (channelNum >= MAX_NUMBER_CHANNELS) 
   {
@@ -372,7 +374,7 @@ void ESP32_PWM_ISR::enable(const unsigned& channelNum)
 
 ///////////////////////////////////////////////////
 
-void ESP32_PWM_ISR::disable(const unsigned& channelNum) 
+void ESP32_PWM_ISR::disable(const uint8_t& channelNum) 
 {
   if (channelNum >= MAX_NUMBER_CHANNELS) 
   {
@@ -427,7 +429,7 @@ void ESP32_PWM_ISR::disableAll()
 
 ///////////////////////////////////////////////////
 
-void ESP32_PWM_ISR::toggle(const unsigned& channelNum) 
+void ESP32_PWM_ISR::toggle(const uint8_t& channelNum) 
 {
   if (channelNum >= MAX_NUMBER_CHANNELS) 
   {
@@ -439,7 +441,7 @@ void ESP32_PWM_ISR::toggle(const unsigned& channelNum)
 
 ///////////////////////////////////////////////////
 
-unsigned ESP32_PWM_ISR::getnumChannels() 
+int8_t ESP32_PWM_ISR::getnumChannels() 
 {
   return numChannels;
 }